horse stable (ha!) trigger and transform (with e )to window top floor left of apartments
The horses now neigh in occordence to their random walk. took a bit to
get them to shut up but its over now. added a similar door trigger to
that window upstairs and now that i understand it more (except for the
text positioning) it can be put in more place. Maybe we should move
control over to the mouse more.....