Added money scripting, taught team about code base
The scripting behind money is ready to be added to. Time for the player to make bank. On top of that, Jack and joe for a little while got to see a bit of the code base and install structure and practices that I'm using. I believe in this so much and i can't wait for others to start making this code base extraordinary. To the future :+1: